refactor: 테스트 코드 디렉토리 구조 정리 및 모듈명 snake_case로 변경
- tests/: 테스트 코드(test_mcp_module.py)를 전용 디렉토리로 이동
- brave-search-module/ → brave_search_module/로 디렉토리명 변경 (PEP8 스타일 적용)
- 경로 변경에 따른 import 경로 수정 및 sys.path 설정 보완
